requirements.txt
Django==5.2.3
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-telegram-bot==22.1
google-generativeai==0.8.5
django-filter==23.5
coreapi==2.3.3
Markdown==3.6
django-cors-headers==4.4.0
python-dotenv==1.0.1
gunicorn==22.0.0 